requirements.txt
pyautogui
opencv-python
numpy
pygetwindow
pywinauto
google-genai
python-dotenv